article popular sort

+++ ms-route/src/main/java/com/thelak/route/article/models/ArticleModel.java
@@ -31,7 +31,7 @@ public class ArticleModel {
 
     Integer rating;
 
-    Long viewsCount;
+    Integer viewsCount;
 
     List<CategoryModel> categories;
